POI：Bug 57915: Fix Dev2Hex for numbers larger than Integer.MAX_VALUE and less than Integer.MIN_VALUE

+++ main/SS/Formula/Functions/Dec2Hex.cs
@@ -127,7 +127,7 @@ namespace NPOI.SS.Formula.Functions
             }
             else
             {
-                hex = String.Format("{0:X}", (int)number1);
+                hex = String.Format("{0:X}", (long)number1);
             }
             if (number1 < 0)
             {
